Refactor market depth project structure
