+++ tests/BinanceBot.Market.Tests/MarketDepthManagerTests.cs
@@ -2,6 +2,7 @@ using System.Threading;
 using System.Threading.Tasks;
 using Binance.Net.Interfaces.Clients;
 using BinanceBot.Market.Core;
+using BinanceBot.Market.Domain;
 using Moq;
 using NLog;
